feat: tty without collapsing
