+++ cmd/stack/deploy.go
@@ -119,7 +119,7 @@ func (c *DeployCommand) RunDeploy(ctx context.Context, s *model.Stack, options *
 	}
 	oktetoLog.Success("Compose '%s' successfully deployed", s.Name)
 
-	if !utils.LoadBoolean(model.OktetoWithinDeployCommandContextEnvVar) || !c.IsInsideDeploy {
+	if !(!utils.LoadBoolean(model.OktetoWithinDeployCommandContextEnvVar) || !c.IsInsideDeploy) {
 		if err := stack.ListEndpoints(ctx, s, ""); err != nil {
 			return err
 		}
